Refactor frontend with modular API and CoreUI layout
