+++ frontend/src/components/QRScanner.js
@@ -1,29 +1,28 @@
-import React, { useEffect, useState } from "react";
-import axios from "axios";
-const API = "http://localhost:3001";
+import React, { useEffect, useState } from 'react';
+import api from '../api';
 
 const QRScanner = ({ sessionId }) => {
   const [qr, setQr] = useState(null);
-  const [status, setStatus] = useState("");
+  const [status, setStatus] = useState('');
   const [info, setInfo] = useState(null);
 
   useEffect(() => {
     if (!sessionId) {
       setQr(null);
-      setStatus("");
+      setStatus('');
       setInfo(null);
       return;
     }
     const interval = setInterval(() => {
-      axios.get(`${API}/sessions/${sessionId}/status`)
+      api.get(`/sessions/${sessionId}/status`)
         .then(res => {
           setStatus(res.data.status);
           setInfo(res.data.info);
         }).catch(() => {
-          setStatus("");
+          setStatus('');
           setInfo(null);
         });
-      axios.get(`${API}/sessions/${sessionId}/qr`)
+      api.get(`/sessions/${sessionId}/qr`)
         .then(res => setQr(res.data.qr))
         .catch(() => setQr(null));
     }, 1500);
@@ -36,11 +35,11 @@ const QRScanner = ({ sessionId }) => {
     <div>
       <h2>Status Session: {sessionId}</h2>
       <div>
-        {status === "connected" ? (
-          <div style={{color:"green",fontWeight:"bold",fontSize:"1.1em"}}>
+        {status === 'connected' ? (
+          <div style={{color:'green',fontWeight:'bold',fontSize:'1.1em'}}>
             Connected ✅<br />
-            <span style={{color:"#0066cc"}}>
-              {info?.wid?.user} {info?.pushname ? `(${info.pushname})` : ""}
+            <span style={{color:'#0066cc'}}>
+              {info?.wid?.user} {info?.pushname ? `(${info.pushname})` : ''}
             </span>
           </div>
         ) : (
@@ -48,7 +47,7 @@ const QRScanner = ({ sessionId }) => {
             {qr ? (
               <>
                 <img src={qr} alt="QR" />
-                <div style={{ color: "#0066cc", marginTop: 6 }}>
+                <div style={{ color: '#0066cc', marginTop: 6 }}>
                   Scan QR di WhatsApp (Perangkat Tertaut)
                 </div>
               </>
